refactor(compile): extract per-receiver branches from EmitGetProperty (#1242)
Slice 15 self-contained, disjoint receiver arms out of the 2,028-line
EmitGetProperty god-method into named Emit<Receiver>GetBranch helpers in a
new sibling partial (RuntimeEmitter.Objects.GetPropertyBranches.cs). Each
helper mirrors the already-factored EmitProxyGetPropertyCheck shape: it emits
the receiver-type test itself, runs the branch body + `ret` on a match, and
otherwise falls through to a caller-supplied `notMatch` label. The dispatch
table in EmitGetProperty now reads as a flat sequence of
`EmitXGetBranch(il, …, notX); MarkLabel(notX)` arms.

Extracted arms: namespace, $Object, Map, Set, Dictionary, $Array, List,
object[], String, $Buffer, $Stats, $TSFunction, $BoundTSFunction, $CJSModule,
$RegExp. EmitGetProperty drops from ~2,028 to ~960 lines.

The interdependent arms deferred in #1141/#1242 stay inline: the
System.Type/class-instance pair (shared base-walk jump target), the three-way
Promise dispatch, the typed-array family, the bound-callable multi-check, and
the primitive-prototype fall-throughs.

Pure mechanical move — no behaviour change. Each arm was extracted
incrementally and verified with `--compile … --verify` plus an output diff
against a pre-refactor compiled baseline.

Gate: xUnit 15404/0, Test262 22/0/4skip (interp + compiled), TSConformance
31/0, IL --verify clean. Standalone constraint preserved (no new metadata
tokens; existing Ldtoken/GetMethodFromHandle late-binding retained).

Part of #1094.
